+++ packages/less-plugin-dls/src/inject.js
@@ -6,7 +6,11 @@ const ENTRY_LESS = path.resolve(__dirname, '../tokens/index.less')
 class Injector {
   process (src, extra) {
     // Don't inject self
-    if (extra.fileInfo.filename.indexOf(SELF_MODULE_PATH) >= 0) {
+    if (
+      path
+        .resolve(process.cwd(), extra.fileInfo.filename)
+        .indexOf(SELF_MODULE_PATH) >= 0
+    ) {
       return src
     }
 
@@ -18,9 +22,9 @@ class Injector {
     if (relative.charAt(0) !== '.') {
       relative = `./${relative}`
     }
-    let injected = `@import "${relative}";\n`
-    let ignored = extra.imports.contentsIgnoredChars
-    let fileInfo = extra.fileInfo
+    const injected = `@import "${relative}";\n`
+    const ignored = extra.imports.contentsIgnoredChars
+    const fileInfo = extra.fileInfo
     ignored[fileInfo.filename] = ignored[fileInfo.filename] || 0
     ignored[fileInfo.filename] += injected.length
     return injected + src